requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
autopep8==2.0.4
cffi==1.15.1
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
Faker==19.5.0
iniconfig==2.0.0
Markdown==3.4.4
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
